+++ BuyRecipes/release-notes.md
@@ -1,4 +1,14 @@
 ﻿[← back to readme](readme.md)
 
 # Release notes
+## 0.3.0
+* Added version checking
+
+## 0.2.0
+* console now outputs weekly recipes each sunday
+
 ## 0.1.0
+* initial cut
+* all access to the mod is via SMAPI console
+* up to 5 recipes are available each week
+* recipes will only be ones you don't know